Create Emailslicer.py
The code above is very simple and easy to understand. I take user input and use the strip function at the same time to remove white space if any. Then I am finding the index of ‘@’ symbol of the user input. Then I store the index into a variable known as domain_name to split the email into two parts; the user name and the domain.